+++ src/util/proximityProportions.ts
@@ -14,6 +14,7 @@ import {
 } from "shared";
 import { forEachPlayer, reducePlayerUnits } from "util/temp";
 import { adjustPlayerGold } from "resources/goldPerSecond";
+import { isPlayingPlayer, isComputer } from "./player";
 
 export const GOLEM_TYPE = FourCC( "ewsp" );
 const STALKER_TYPE = FourCC( "nfel" );
@@ -28,17 +29,6 @@ type Amount = {
     lumber?: number;
     experience?: number;
 }
-
-const WIDTH = 5376 * 2;
-
-const SHEEP_DISTANCE_FACTOR = 1 / 27;
-const SHEEP_DENOM = ( WIDTH ** 2 * 2 - 128 ) ** SHEEP_DISTANCE_FACTOR;
-
-const WISP_DISTANCE_FACTOR = 1 / 27;
-const WISP_DENOM = ( ( WIDTH / 10 ) ** 2 * 2 - 128 ) ** WISP_DISTANCE_FACTOR;
-
-const WOLF_DISTANCE_FACTOR = 1 / 9;
-const WOLF_DENOM = ( WIDTH ** 2 * 2 - 128 ) ** WOLF_DISTANCE_FACTOR;
 
 const UNIT_FACTORS: Record<number, number> = {
 	[ WOLF_TYPE ]: 1,
@@ -61,12 +51,33 @@ const unitFactor = ( unit: unit ): number => {
 
 };
 
+/**
+ * Returns an array where all values sum to 1.
+ */
 export const normalize = ( arr: number[] ): number[] => {
 
 	const sum = arr.reduce( ( sum, v ) => sum + v, 0 );
 	return arr.map( v => v / sum );
 
 };
+
+/**
+ * Returns a sigmoid function using the passed constants.
+ * @param v Controls the impact of inflection.
+ * @param b Linear transformer on `t`.
+ * @param c Controls how far t=0 is along the curve.
+ */
+const sigmoid = ( v: number, b: number, c: number ):
+( ( t: number ) => number ) => {
+
+	const cTetrated = c ** c;
+	const k = ( cTetrated + 1 ) ** v;
+	return ( t: number ): number =>
+		k / ( cTetrated + Math.exp( b * t ) ) ** v;
+
+};
+
+const sheepSigmoid = sigmoid( 0.04, 0.03, 4 );
 
 const sheepProximityProportions = (
 	{ x, y }: Point,
@@ -78,10 +89,15 @@ const sheepProximityProportions = (
 	let proportions: number[] = [];
 	forEachPlayer( p => {
 
-		if ( ! IsPlayerInForce( p, sheepTeam ) ) return;
+		if (
+			! IsPlayerInForce( p, sheepTeam ) ||
+			! isPlayingPlayer( p ) && ! isComputer( p )
+		) return;
 
-		const distanceSquared = ( GetUnitX( sheeps[ GetPlayerId( p ) ] ) - x ) ** 2 + ( GetUnitY( sheeps[ GetPlayerId( p ) ] ) - y ) ** 2;
-		const proportion = 1 - Math.max( distanceSquared - 128, 0 ) ** SHEEP_DISTANCE_FACTOR / SHEEP_DENOM;
+		const xDelta = GetUnitX( sheeps[ GetPlayerId( p ) ] ) - x;
+		const yDelta = GetUnitY( sheeps[ GetPlayerId( p ) ] ) - y;
+		const distance = ( xDelta ** 2 + yDelta ** 2 ) ** 0.5;
+		const proportion = sheepSigmoid( distance );
 
 		proportions.push( proportion );
 		players.push( p );
@@ -119,6 +135,8 @@ const sheepProximityProportions = (
 	return proportionAmounts;
 
 };
+
+const wispSigmoid = sigmoid( 0.08, 0.08, 8 );
 
 const wispProximityProportions = (
 	{ x, y }: Point,
@@ -130,10 +148,15 @@ const wispProximityProportions = (
 	let proportions: number[] = [];
 	forEachPlayer( p => {
 
-		if ( ! IsPlayerInForce( p, wispTeam ) ) return;
+		if (
+			! IsPlayerInForce( p, wispTeam ) ||
+			! isPlayingPlayer( p ) && ! isComputer( p )
+		) return;
 
-		const distanceSquared = ( GetUnitX( wisps[ GetPlayerId( p ) ] ) - x ) ** 2 + ( GetUnitY( wisps[ GetPlayerId( p ) ] ) - y ) ** 2;
-		const proportion = 1 - Math.max( distanceSquared - 128, 0 ) ** WISP_DISTANCE_FACTOR / WISP_DENOM;
+		const xDelta = GetUnitX( wisps[ GetPlayerId( p ) ] ) - x;
+		const yDelta = GetUnitY( wisps[ GetPlayerId( p ) ] ) - y;
+		const distance = ( xDelta ** 2 + yDelta ** 2 ) ** 0.5;
+		const proportion = wispSigmoid( distance );
 
 		proportions.push( proportion );
 		players.push( p );
@@ -172,10 +195,14 @@ const wispProximityProportions = (
 
 };
 
+const wolfSheepKillSigmoid = sigmoid( 0.01, 0.02, 4 );
+const wolfOtherKillSigmoid = sigmoid( 0.04, 0.04, 4 );
+
 // wolves shouldn't have too many units, so fine to iterate over them all
 const wolfProximityProportions = (
 	{ x, y }: Point,
 	amounts: Amount,
+	killType: "sheep" | "other" = "other",
 ): Map<player, Amount> => {
 
 	// calculate proportions, which is the max of the players' units' proportion
@@ -187,8 +214,14 @@ const wolfProximityProportions = (
 
 		const proportion = reducePlayerUnits( p, ( max, unit ) => {
 
-			const distanceSquared = ( GetUnitX( unit ) - x ) ** 2 + ( GetUnitY( unit ) - y ) ** 2;
-			const proportion = ( 1 - Math.max( distanceSquared - 128, 0 ) ** WOLF_DISTANCE_FACTOR / WOLF_DENOM ) * unitFactor( unit );
+			const xDelta = GetUnitX( unit ) - x;
+			const yDelta = GetUnitY( unit ) - y;
+			const distance = ( xDelta ** 2 + yDelta ** 2 ) ** 0.5;
+			const proportion = ( killType === "sheep" ?
+				wolfSheepKillSigmoid :
+				wolfOtherKillSigmoid
+			)( distance ) * unitFactor( unit );
+
 			if ( proportion > max )	return proportion;
 			return max;
 
@@ -235,10 +268,11 @@ export const proximityProportions = (
 	origin: Point,
 	amounts: Amount,
 	killer: player,
+	killType: "sheep" | "other" = "other",
 ): Map<player, Amount> => {
 
 	if ( IsPlayerInForce( killer, wolfTeam ) )
-		return wolfProximityProportions( origin, amounts );
+		return wolfProximityProportions( origin, amounts, killType );
 
 	if ( IsPlayerInForce( killer, sheepTeam ) )
 		return sheepProximityProportions( origin, amounts );
@@ -251,9 +285,10 @@ export const awardBounty = (
 	origin: Point,
 	amounts: Amount,
 	killer: player,
+	killType: "sheep" | "other" = "other",
 ): void => {
 
-	const bounties = proximityProportions( origin, amounts, killer );
+	const bounties = proximityProportions( origin, amounts, killer, killType );
 
 	bounties.forEach( ( bounty, player ) => {