+++ CHANGELOG.md
@@ -10,8 +10,12 @@
 - "-wolf gold" and "-destroy all farms" prefer host
 - Added -gps
 
+## Bug fixes
+- Fixed a bug with calculating bounties for saves.
+
 ## Refactors
 - Rewrote w3mmd, aiming for less desyncs + type safety
+- Switched bounty proximities to a sigmoid.
 
 # Version 10a
 ## Features